pyproject.toml
[tool.poetry]
name = "gfycat"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
loguru = "^0.6.0"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
isort = "^5.10.1"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py38", "py39", "py310"]
line-length = 120